app/requirements.txt
ruff==0.4.1
flask==3.0.0
pytest==8.1.0
psycopg2-binary==2.9.9